+++ propagation/baggage_test.go
@@ -47,7 +47,7 @@ func (m member) Member(t *testing.T) baggage.Member {
 		}
 		props = append(props, p)
 	}
-	bMember, err := baggage.NewMember(m.Key, url.QueryEscape(m.Value), props...)
+	bMember, err := baggage.NewMember(m.Key, url.PathEscape(m.Value), props...)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -244,6 +244,55 @@ func TestInjectBaggageToHTTPReq(t *testing.T) {
 
 			got := strings.Split(req.Header.Get("baggage"), ",")
 			assert.ElementsMatch(t, tt.wantInHeader, got)
+		})
+	}
+}
+
+func TestBaggageInjectExtractRoundtrip(t *testing.T) {
+	propagator := propagation.Baggage{}
+	tests := []struct {
+		name string
+		mems members
+	}{
+		{
+			name: "two simple values",
+			mems: members{
+				{Key: "key1", Value: "val1"},
+				{Key: "key2", Value: "val2"},
+			},
+		},
+		{
+			name: "values with escaped chars",
+			mems: members{
+				{Key: "key1", Value: "val3=4"},
+				{Key: "key2", Value: "mess,me%up"},
+			},
+		},
+		{
+			name: "with properties",
+			mems: members{
+				{Key: "key1", Value: "val1"},
+				{
+					Key:   "key2",
+					Value: "val2",
+					Properties: []property{
+						{Key: "prop", Value: "1"},
+					},
+				},
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			b := tt.mems.Baggage(t)
+			req, _ := http.NewRequest("GET", "http://example.com", nil)
+			ctx := baggage.ContextWithBaggage(context.Background(), b)
+			propagator.Inject(ctx, propagation.HeaderCarrier(req.Header))
+
+			ctx = propagator.Extract(context.Background(), propagation.HeaderCarrier(req.Header))
+			got := baggage.FromContext(ctx)
+
+			assert.Equal(t, b, got)
 		})
 	}
 }